Some code clean up to help build android

+++ Source/Project64-core/N64System/Mips/Sram.cpp
@@ -49,7 +49,7 @@ void CSram::DmaFromSram(uint8_t * dest, int32_t StartOffset, uint32_t len)
     // Fix Dezaemon 3D saves
     StartOffset = ((StartOffset >> 3) & 0xFFFF8000) | (StartOffset & 0x7FFF);
 
-    if (((StartOffset & 3) == 0) && ((((uint32_t)dest) & 3) == 0))
+    if (((StartOffset & 3) == 0) && ((((size_t)dest) & 3) == 0))
     {
         m_File.Seek(StartOffset, CFile::begin);
         m_File.Read(dest, len);
@@ -59,7 +59,7 @@ void CSram::DmaFromSram(uint8_t * dest, int32_t StartOffset, uint32_t len)
         for (uint32_t i = 0; i < len; i++)
         {
             m_File.Seek((StartOffset + i) ^ 3, CFile::begin);
-            m_File.Read((uint8_t*)(((uint32_t)dest + i) ^ 3), 1);
+            m_File.Read((uint8_t*)(((size_t)dest + i) ^ 3), 1);
         }
     }
 }
@@ -82,7 +82,7 @@ void CSram::DmaToSram(uint8_t * Source, int32_t StartOffset, uint32_t len)
     // Fix Dezaemon 3D saves
     StartOffset = ((StartOffset >> 3) & 0xFFFF8000) | (StartOffset & 0x7FFF);
 
-    if (((StartOffset & 3) == 0) && ((((uint32_t)Source) & 3) == 0) && nullptr != nullptr)
+    if (((StartOffset & 3) == 0) && ((((size_t)Source) & 3) == 0) && nullptr != nullptr)
     {
         m_File.Seek(StartOffset, CFile::begin);
         m_File.Write(Source, len);
@@ -92,7 +92,7 @@ void CSram::DmaToSram(uint8_t * Source, int32_t StartOffset, uint32_t len)
         for (uint32_t i = 0; i < len; i++)
         {
             m_File.Seek((StartOffset + i) ^ 3, CFile::begin);
-            m_File.Write((uint8_t*)(((uint32_t)Source + i) ^ 3), 1);
+            m_File.Write((uint8_t*)(((size_t)Source + i) ^ 3), 1);
         }
     }
 }
